# docs/swagger.yaml
basePath: /api/v1
definitions:
  barber-booking-system_internal_middleware.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      error:
        type: string
      message:
        type: string
    type: object
...
  barber-booking-system_internal_services.CreateNotificationRequest:
    properties:
      channels:
        description: Optional fields
        items:
          type: string
        type: array
      data:
        additionalProperties: true
        type: object
      expires_at:
        type: string
      message:
        maxLength: 2000
        minLength: 1
        type: string
      priority:
        enum:
        - low
        - normal
        - high
        - urgent
        type: string
      related_entity_id:
        type: integer
      related_entity_type:
        type: string
      scheduled_for:
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
      type:
        type: string
      user_id:
        type: integer
    required:
    - message
    - title
    - type
    - user_id
    type: object
...
  barber-booking-system_internal_services.SendBookingNotificationRequest:
    properties:
      booking_id:
        type: integer
      custom_message:
        type: string
      notification_type:
        type: string
    required:
    - booking_id
    - notification_type
    type: object
...
  internal_handlers.SuccessResponse:
    description: Standard success response wrapper
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      meta: {}
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@barbershop.com
    name: API Support
  description: |-
    A production-ready RESTful API for barbershop booking system.
    Features include: User authentication, Barber management, Service catalog, Booking system, Review & Rating system, and Notification system.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Barbershop Booking API
  version: 1.0.0
paths:
...
  /api/v1/notifications:
    get:
      consumes:
      - application/json
      description: Get all notifications for the authenticated user
      parameters:
      - description: Filter by notification type
        in: query
        name: type
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by priority
        in: query
        name: priority
        type: string
      - description: Filter by read status
        in: query
        name: is_read
        type: boolean
      - description: Filter by unread status
        in: query
        name: is_unread
        type: boolean
      - default: created_at
        description: Sort by field
        in: query
        name: sort_by
        type: string
      - default: DESC
        description: Sort order (ASC/DESC)
        in: query
        name: order
        type: string
      - default: 50
        description: Limit results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my notifications
      tags:
      - notifications
    post:
      consumes:
      - application/json
      description: Create a new notification for a user (admin only)
      parameters:
      - description: Notification data
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/barber-booking-system_internal_services.CreateNotificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a notification
      tags:
      - notifications
  /api/v1/notifications/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific notification
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a notification
      tags:
      - notifications
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific notification
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get notification by ID
      tags:
      - notifications
  /api/v1/notifications/{id}/read:
    patch:
      consumes:
      - application/json
      description: Mark a specific notification as read
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - notifications
  /api/v1/notifications/{id}/webhook:
    post:
      consumes:
      - application/json
      description: Webhook endpoint for push notification delivery callbacks
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      - description: Delivery status (delivered/failed)
        in: query
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      summary: Notification delivery webhook
      tags:
      - notifications
  /api/v1/notifications/booking:
    post:
      consumes:
      - application/json
      description: Send a notification related to a booking
      parameters:
      - description: Booking notification request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/barber-booking-system_internal_services.SendBookingNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send booking notification
      tags:
      - notifications
  /api/v1/notifications/read-all:
    patch:
      consumes:
      - application/json
      description: Mark all notifications for the authenticated user as read
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark all notifications as read
      tags:
      - notifications
  /api/v1/notifications/stats:
    get:
      consumes:
      - application/json
      description: Get notification statistics for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get notification statistics
      tags:
      - notifications
  /api/v1/notifications/unread:
    get:
      consumes:
      - application/json
      description: Get unread notifications for the authenticated user
      parameters:
      - default: 20
        description: Limit results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get unread notifications
      tags:
      - notifications
  /api/v1/notifications/unread/count:
    get:
      consumes:
      - application/json
      description: Get the count of unread notifications for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get unread notification count
      tags:
      - notifications
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...